warmthcompetence: Warmth and Competence Detectors

Detects perceptions of warmth and competence in American English self-presentational language. Using pre-trained enet regression models, this package provides a numerical representation of warmth and competence perceptions.

See also

Author

Maintainer: Bushra Guenoun bushraguenoun@gmail.com (ORCID)

Authors:

Julian Zlatev (ORCID)

Other contributors:

Noah Greifer ngreifer@iq.harvard.edu (ORCID) [contributor]